Implemented Search and Filter functionalities with search view in MapFragment.kt and RestaurantListFragment.kt. User is now able to search for and filter through list of restaurants in RestaurantListFragment.kt or search and filter through restaurant markers in MapFragment.kt. I implemented the functionality in the following way: - call "setOnTextChanged" listener to searchview - Retrieve data from viewmodel - Filter data from viewmodel -Use filtered data to update/populate either markers in the MapFragment.kt or the adapter being initialized in the RestaurantListFragment.kt
Also, I implemented functionality to add each user to each restaurant in firestore database so that the number of users planning to eat at each restaurant is displayed in a textview in the RestaurantListFragment.kt.
Furthermore, I also made sure to add a geolocalization button in the MapFragment.kt which centers the map.
I also implemented the "Like" functionality for the RestaurantDetailsActivity.kt ; user can now "Like" and "Unlike" any restaurant from the RestaurantDetailsActivity.kt. Once the user has "Liked" a restaurant, the Star icon changes colors from black to mint (primaryDark in colors.xml file) and "Like" textview text changes to "Liked". Value is saved to shared preferences, retrieved each time user goes into RestaurantDetailsActivity.kt and removed whenever user "Unlikes" the restaurant.
Lastly, I added an onClickListener to the "Your Lunch" textview located in the Nav Drawer. Once user clicks the textview, if they currently have a restaurant selected , the data for that restaurant will be retrieved from firestore and passed to the RestaurantDetailsActivity.kt to view their restaurant. Otherwise, if the user does not currently have a selected restaurant when clicking the "Your Lunch" textview, a toast displays to let them know that they haven't selected a restaurant yet.