Implemented "Visit Website" and "Call" functionalities in RestaurantDetailsActivity.kt ; Added "GetPlaceDetails" @GET function to ApiInterface.kt with parameters to fetch the website url and phone number of current restaurant. I added the api call to my ViewModel and attached the livedata observers to the RestaurantDetailsActivity.kt where if the api call is successfull, onClickListeners are set to Call Icon and Website Icon so user can visit the restaurant's website and call the restaurant as well.
Also, I made a couple of changes to the UI as well. I added a transparent status bar to the RestaurantDetailsActivity.kt and changed the color of my imageview that holds the Call Icon, Like Icon and Website Icon to Pink for a more appealing look as it seperates the Recycleriew and Icons in an aesthetically pleasing way.